% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_vcf_missing_data.R
\name{assess_vcf_missing_data}
\alias{assess_vcf_missing_data}
\title{assess_vcf_missing_data}
\usage{
assess_vcf_missing_data(
  vcf,
  data_path,
  res_path,
  project,
  postfix,
  fltr,
  species,
  strt = "strata"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{data_path}{-> path to data (where strata is located)}

\item{res_path}{-> path to results (directory for output dataframe and plot)}

\item{project}{-> project base name (character)}

\item{postfix}{-> VCF extraction method (character)}

\item{fltr}{-> VCF filtration method (character)}

\item{species}{-> species name for plot (character)}

\item{strt}{-> file with 2+ columns, one id column and one pop column (tsv file, header = TRUE)}
}
\value{
dataframe and plot of missing data for each individual in VCF
}
\description{
quantifying missing data of all samples in VCF

inspired by https://grunwaldlab.github.io/Population_Genetics_in_R/qc.html
}
\details{
This function generates a dataframe of absolute and relative missing data per sample
This function plots relative missing data per individual, with individuals sorted by group
The postfix and fltr parameters can have empty values ("")
}
\examples{
assess_vcf_missing_data(vcf = my_vcf, data_path = data_path, res_path = res_path, project = project_name, postfix = extraction_info, fltr = filter_parms, species = species_name, strt = "strata")
assess_vcf_missing_data(my_vcf, data_path, res_path, project_name, extraction_info, filter_parms, species_name, strt = "strata")
assess_vcf_missing_data(my_vcf, data_path, res_path, project_name, postfix, fltr, species)

}
\author{
Tomas Hrbek July 2022
}
