% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcf_group_info.R
\name{get_vcf_group_info}
\alias{get_vcf_group_info}
\title{get_vcf_group_info}
\usage{
get_vcf_group_info(vcf, data_path, strt = "strata", grps = "groups")
}
\arguments{
\item{vcf}{-> vcfR object}

\item{data_path}{-> path to data (where strata and groups are located)}

\item{strt}{-> file with 2+ columns, one id column and one pop column (tsv file, header = TRUE)}

\item{grps}{-> one column file listing groups to include, column name pop (header = TRUE)}
}
\value{
list of factors of sample-to-group assignments and groups
}
\description{
associate samples with group info based on "strata" file

"strata" is a tsv file with 2+ columns, with 2 columns name id and pop
}
\details{
This function creates list of factors of sample-to-group assignments and groups based on a strata and group file
}
\examples{
get_vcf_group_info(vcf = my_vcf, data_path = data_path, strt = "strata", grps = "groups")
get_vcf_group_info(my_vcf, data_path, strt = "strata", grps = "groups")
get_vcf_group_info(my_vcf, data_path)

}
\author{
Tomas Hrbek July 2022
}
