% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2apparent.R
\name{vcf2apparent}
\alias{vcf2apparent}
\title{vcf2apparent}
\usage{
vcf2apparent(
  vcf,
  ind_pop,
  keep_pop,
  key = "All",
  inc_missing = TRUE,
  out_file = "apparent_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Apparent infile (factor)}

\item{key}{-> relationship type (All, Pa, Mo, Fa, Off); default All (character)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Apparent infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Apparent infile
}
\details{
This function converts the vcfR object to a Apparent formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2apparent(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, key = key, inc_missing = TRUE, out_file = "Genepop_infile.gen")
vcf2apparent(my_vcf, ind_pop, keepers, out_file = "Apparent_infile.txt")
vcf2apparent(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek February 2023
}
