% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2arlequin.R
\name{vcf2arlequin}
\alias{vcf2arlequin}
\title{vcf2arlequin}
\usage{
vcf2arlequin(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "arlequin.arp"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Arlequin infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Arlequin infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Arlequin infile

in part based on vcfR2migrate function (vcfR package)
}
\details{
This function converts the vcfR object to a Arlequin formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2arlequin(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "Arlequin_infile.arp")
vcf2arlequin(my_vcf, ind_pop, keepers, out_file = "Arlequin_infile.arp")
vcf2arlequin(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek December 2020
}
