% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2bayescan.R
\name{vcf2bayescan}
\alias{vcf2bayescan}
\title{vcf2bayescan}
\usage{
vcf2bayescan(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "bayescan_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Bayescan infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Bayescan infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Bayescan infile

in part based on vcfR2migrate function (vcfR package)
}
\details{
This function converts the vcfR object to a Bayescan formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2bayescan(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "Bayescan_infile.txt")
vcf2bayescan(my_vcf, ind_pop, keepers, out_file = "Bayescan_infile.txt")
vcf2bayescan(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek November 2022
}
