% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{vcf2df}
\usage{
vcf2df(
  vcf,
  ind_pop,
  keep_pop,
  alleles = TRUE,
  inc_missing = TRUE,
  out_file = "dataframe_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in DataFrame infile (factor)}

\item{alleles}{-> record data as alleles (GATC) or SNPs (012) (logical)}

\item{inc_missing}{-> include missing data (logical)}
}
\value{
nothing
}
\description{
converts vcfR format data to DataFrame infile

adds last column of an individual's group origin

motivated to provide a dataframe for plotting PCA results

for use by modified PCA function from adegenet
}
\details{
This function converts the vcfR object to a DataFrame formatted input file
This function adds group (population) membership as last column
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2df(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, alleles = TRUE, inc_missing = TRUE, out_file = "dataframe_infile.txt")
vcf2df(my_vcf, ind_pop, keepers, out_file = "dataframe_infile.txt")
vcf2df(my_vcf, ind_pop, keepers)
}
\author{
Tomas Hrbek December 2022
}
