% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2eigenstrat.R
\name{vcf2eigenstrat}
\alias{vcf2eigenstrat}
\title{vcf2eigenstrat}
\usage{
vcf2eigenstrat(
  vcf,
  ind_pop,
  keep_pop,
  sex = "U",
  rel_pos = 0,
  inc_missing = TRUE,
  out_file = "eigenstrat_infile"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Eigenstrat infiles (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Eigenstrat infiles)}
}
\value{
nothing
}
\description{
converts vcfR format data to Eigenstrat infiles

in part based on vcfR2migrate function (vcfR package)
}
\details{
This function converts the vcfR object to a Eigenstrat formatted input files
When list of sexes is not provided, lists all individuals as unknown
When relative position on chromosome (cM distance or similar) is not provides, list as 0
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2eigenstrat(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, sex = list_of_sex, rel_pos = marker_cM_map, inc_missing = TRUE, out_file = "Eigenstrat")
vcf2eigenstrat(my_vcf, ind_pop, keepers, out_file = "Eigenstrat")
vcf2eigenstrat(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek December 2022
}
