% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2fasta.R
\name{vcf2fasta}
\alias{vcf2fasta}
\title{vcf2fasta}
\usage{
vcf2fasta(
  vcf,
  ind_pop,
  keep_pop,
  interleaved = FALSE,
  inc_missing = TRUE,
  out_file = "fasta.fas"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Fasta infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Fasta infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Fasta format infile
}
\details{
This function converts the vcfR object to a Fasta formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2fasta(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, interleaved = FALSE, inc_missing = TRUE, out_file = "Fasta_infile.fas")
vcf2fasta(my_vcf, ind_pop, keepers, out_file = "Fasta_infile.fas")
vcf2fasta(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek August 2022
}
