% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2genepop.R
\name{vcf2genepop}
\alias{vcf2genepop}
\title{vcf2genepop}
\usage{
vcf2genepop(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "genepop_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Genepop infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Genepop infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Genepop infile

includes population names that can be ready by read_genepop {}

in part based on vcfR2migrate function (vcfR package)
}
\details{
This function converts the vcfR object to a Genepop formatted input file
This function labels populations. To read labeled populations use "read_genepop" function
The function will remove indels, and multiallelic loci, and optionally loci with missing data
01, 02, 03, 04 is 'A', 'C', 'G', 'T'
}
\examples{
vcf2genepop(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "Genepop_infile.gen")
vcf2genepop(my_vcf, ind_pop, keepers, out_file = "Genepop_infile.gen")
vcf2genepop(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek January 2021
}
