% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2genlight.R
\name{vcf2genlight}
\alias{vcf2genlight}
\title{vcf2genlight}
\usage{
vcf2genlight(
  vcf,
  ind_pop,
  keep_pop,
  ploidy = 2,
  inc_missing = TRUE,
  save = FALSE,
  out_file = "genlight_infile"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Genlight infile (factor)}

\item{ploidy}{-> ploidy level (default 2)}

\item{inc_missing}{-> include missing data (logical)}

\item{save}{-> save to file (logical - default TRUE) or just return Genlight object}
}
\value{
Genlight object
}
\description{
converts vcfR format data to Genlight infile

in part based on vcfR2genlight function (vcfR package)
}
\details{
This function converts the vcfR object to a Genlight formatted input file
This function labels populations.
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2genlight(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, ploidy = 2, inc_missing = TRUE, out_file = "Genlight_infile")
vcf2genlight(my_vcf, ind_pop, keepers, out_file = "Genlight_infile")
vcf2genlight(my_vcf, ind_pop, keepers)
}
\author{
Tomas Hrbek December 2022
}
