% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2migrate (copy).R, R/vcf2migrate.R
\name{vcf2migrate}
\alias{vcf2migrate}
\title{vcf2migrate}
\usage{
vcf2migrate(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "migrateN_infile.txt",
  method = "N"
)

vcf2migrate(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "migrateN_infile.txt",
  method = "N"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in MigrateN infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (MigrateN infile)}

\item{method}{-> classic or het format}
}
\value{
nothing

nothing
}
\description{
converts vcfR format data to MigrateN infile

adapted vcfR2migrate function (vcfR package) to allow for inclusion of missing data in Migrate output

converts vcfR format data to MigrateN infile

adapted vcfR2migrate function (vcfR package) to allow for inclusion of missing data in Migrate output
}
\details{
This function converts the vcfR object to a MigrateN formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data

This function converts the vcfR object to a MigrateN formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2migrate(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "MigrateN_infile.txt", method = "N")
vcf2migrate(my_vcf, ind_pop, keepers, out_file = "MigrateN_infile.txt")
vcf2migrate(my_vcf, ind_pop, keepers)

vcf2migrate(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "MigrateN_infile.txt")
vcf2migrate(my_vcf, ind_pop, keepers, out_file = "MigrateN_infile.txt")
vcf2migrate(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek August 2020
}
