% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2nexus.R
\name{vcf2nexus}
\alias{vcf2nexus}
\title{vcf2nexus}
\usage{
vcf2nexus(vcf, ind_pop, keep_pop, inc_missing = TRUE, out_file = "nexus.nex")
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Nexus infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Nexus infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Nexus format infile
}
\details{
This function converts the vcfR object to a Nexus formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2nexus(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "Nexus_infile.nex")
vcf2nexus(my_vcf, ind_pop, keepers, out_file = "Nexus_infile.nex")
vcf2nexus(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek February 2021
}
