% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2related.R
\name{vcf2related}
\alias{vcf2related}
\title{vcf2related}
\usage{
vcf2related(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "related_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in Related infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (Related infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to Related infile

in part based on vcfR2genepop function (vcfR package)
}
\details{
This function converts the vcfR object to a Related formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
'A', 'C', 'G', 'T' is 01, 02, 03, 04
}
\examples{
vcf2related(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "Genepop_infile.gen")
vcf2related(my_vcf, ind_pop, keepers, out_file = "Related_infile.txt")
vcf2related(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek April 2023
}
