% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2snapp.R
\name{vcf2snapp}
\alias{vcf2snapp}
\title{vcf2snapp}
\usage{
vcf2snapp(vcf, ind_pop, keep_pop, inc_missing = TRUE, out_file = "snapp.nex")
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in SNAPP infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (SNAPP infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to SNAPP (Nexus) format infile
}
\details{
This function converts the vcfR object to a SNAPP (Nexus) formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2snapp(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "SNAPP_infile.nex")
vcf2snapp(my_vcf, ind_pop, keepers, out_file = "SNAPP_infile.nex")
vcf2snapp(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek December 2020
}
