% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2treemix.R
\name{vcf2treemix}
\alias{vcf2treemix}
\title{vcf2treemix}
\usage{
vcf2treemix(
  vcf,
  ind_pop,
  keep_pop,
  inc_missing = TRUE,
  out_file = "treemix_infile.txt"
)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> population assignment of individuals in vcf (factor)}

\item{keep_pop}{-> population(s) of interest to include in TreeMix infile (factor)}

\item{inc_missing}{-> include missing data (logical)}

\item{out_file}{-> name of file to output (TreeMix infile)}
}
\value{
nothing
}
\description{
converts vcfR format data to TreeMix infile

in part based on vcfR2migrate function (vcfR package)
}
\details{
This function converts the vcfR object to a TreeMix formatted input file
The function will remove indels, and multiallelic loci, and optionally loci with missing data
}
\examples{
vcf2treemix(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, inc_missing = TRUE, out_file = "TreeMix_infile.txt")
vcf2treemix(my_vcf, ind_pop, keepers, out_file = "TreeMix_infile.txt")
vcf2treemix(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek November 2022
}
