% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_add_indivs.R
\name{vcf_add_indivs}
\alias{vcf_add_indivs}
\title{vcf_add_indivs}
\usage{
vcf_add_indivs(vcf, vcf1, indiv, whitelist = TRUE)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{vcf1}{-> vcfR object}

\item{indiv}{-> individuals to add into vcf from vcf1 (factor)}
}
\value{
augmented vcfR object
}
\description{
add individuals to vcf
}
\details{
This function adds individuals from one vcfR object to another vcfR object, returning new vcfR object
}
\examples{
vcf_add_indivs(vcf = my_vcf, vcf1 = other_vcf, indiv = indivs_to_add, whitelist = TRUE)
vcf_add_indivs(vcf = my_vcf, vcf1 = other_vcf, indiv = indivs_to_not_add, whitelist = FALSE)
vcf_add_indivs(my_vcf, other_vcf, indivs_to_add)

}
\author{
Tomas Hrbek October 2023
}
