% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_extract_indivs.R
\name{vcf_extract_indivs}
\alias{vcf_extract_indivs}
\title{vcf_extract_indivs}
\usage{
vcf_extract_indivs(vcf, indiv, whitelist = TRUE, f_invar = TRUE)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{indiv}{-> individuals to retain/drop in vcf (factor)}
}
\value{
extracted vcfR object
}
\description{
extract vcfR format data by individuals
}
\details{
This function extracts the vcfR object by individuals, returning new vcfR object
}
\examples{
vcf_extract_indivs(vcf = my_vcf, indiv = indivs_to_keep, whitelist = TRUE)
vcf_extract_indivs(vcf = my_vcf, indiv = indivs_to_drop, whitelist = FALSE)
vcf_extract_indivs(my_vcf, indivs_to_keep)

}
\author{
Tomas Hrbek December 2020
}
