% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_extract_pops.R
\name{vcf_extract_pops}
\alias{vcf_extract_pops}
\title{vcf_extract_pops}
\usage{
vcf_extract_pops(vcf, ind_pop, keep_pop, whitelist = TRUE, f_invar = TRUE)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{ind_pop}{-> group assignment of individuals in vcf (factor)}

\item{keep_pop}{-> group(s) of interest to include/exclude in vcf infile (factor)}
}
\value{
extracted vcfR object
}
\description{
extract vcfR format data by group
}
\details{
This function extracts the vcfR object by group(s) (individuals
assigned to one or more group), returning new vcfR object
}
\examples{
vcf_extract_pops(vcf = my_vcf, ind_pop = ind_pop, keep_pop = keepers, whitelist = TRUE)
vcf_extract_pops(vcf = my_vcf, ind_pop = ind_pop, keep_pop = non_keepers, whitelist = FALSE)
vcf_extract_pops(my_vcf, ind_pop, keepers)

}
\author{
Tomas Hrbek December 2020
}
