% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_hets.R
\name{vcf_filter_hets}
\alias{vcf_filter_hets}
\title{vcf_filter_hets}
\usage{
vcf_filter_hets(vcf, hets = 0.4)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
remove loci above het threshold from vcfR format data

high hets are indicative of paralogs
}
\details{
This function removes loci above het threshold from the vcfR object
}
\examples{
vcf_filter_hets(vcf = my_vcf, hets = .4)
vcf_filter_hets(my_vcf)

}
\author{
Tomas Hrbek October 2022
}
