% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_invariant.R
\name{vcf_filter_invariant}
\alias{vcf_filter_invariant}
\title{vcf_filter_invariant}
\usage{
vcf_filter_invariant(vcf)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
remove invariant loci from vcfR format data
}
\details{
This function removes invariant loci from the vcfR object
This might be desirable after subsetting the vcf by individuals
}
\examples{
vcf_filter_invariant(vcf = my_vcf)
vcf_filter_invariant(my_vcf)

}
\author{
Tomas Hrbek February 2022
}
