% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_maf.R
\name{vcf_filter_maf}
\alias{vcf_filter_maf}
\title{vcf_filter_maf}
\usage{
vcf_filter_maf(vcf, maf = 0.05)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
remove loci below MAF threshold from vcfR format data
}
\details{
This function removes loci below MAF threshold from the vcfR object
}
\examples{
vcf_filter_maf(vcf = my_vcf, maf = .05)
vcf_filter_maf(my_vcf)

}
\author{
Tomas Hrbek October 2022
}
