% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_missing_indivs.R
\name{vcf_filter_missing_indivs}
\alias{vcf_filter_missing_indivs}
\title{vcf_filter_missing_indivs}
\usage{
vcf_filter_missing_indivs(vcf, p_miss, f_invar = TRUE)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{p_miss}{-> max missing data per locus as decimal (numeric)}
}
\value{
subsetted vcfR object
}
\description{
remove from vcfR format data indivs with >\% missing data
}
\details{
This function subsets the vcfR object by \% missing data within
and individual, returning new vcfR object
}
\examples{
vcf_filter_missing_indivs(vcf = my_vcf, p_miss = p_miss)
vcf_filter_missing_indivs(my_vcf, p_miss)

}
\author{
Tomas Hrbek October 2022
}
