% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_missingness.R
\name{vcf_filter_missingness}
\alias{vcf_filter_missingness}
\title{vcf_filter_missingness}
\usage{
vcf_filter_missingness(vcf, p_miss)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{p_miss}{-> max missing data per locus as decimal (numeric)}
}
\value{
subsetted vcfR object
}
\description{
subsets vcfR format data by \% missing data
}
\details{
This function subsets the vcfR object by \% missing data, returning new vcfR object
}
\examples{
vcf_filter_missingness(vcf = my_vcf, p_miss = p_miss)
vcf_filter_missingness(my_vcf, p_miss)

}
\author{
Tomas Hrbek February 2022
}
