% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_multiSNP.R
\name{vcf_filter_multiSNP}
\alias{vcf_filter_multiSNP}
\title{vcf_filter_multiSNP}
\usage{
vcf_filter_multiSNP(vcf, minS = 2, maxS = 5)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
subsets vcfR format data keeping only loci with 2+ SNPs
}
\details{
This function subsets the vcfR object keeping only loci with between min and max # of SNPs
per locus, returning new vcfR object
default min = 2 and max = 5 SNPs per locus
Recommended as input for fineRADstructure analyses
}
\examples{
vcf_filter_multiSNP(vcf = my_vcf, minS = 2, maxS = 5)
vcf_filter_multiSNP(vcf = my_vcf)
vcf_filter_multiSNP(my_vcf)

}
\author{
Tomas Hrbek February 2022
}
