% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_oneSNP.R
\name{vcf_filter_oneSNP}
\alias{vcf_filter_oneSNP}
\title{vcf_filter_oneSNP}
\usage{
vcf_filter_oneSNP(vcf)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
subsets vcfR format data keeping only 1 SNP per locus
}
\details{
This function subsets the vcfR object keeping only 1 SNP per locus, returning new vcfR object
The first SNP independent of quality is taken (may mofify this in future)
}
\examples{
vcf_filter_oneSNP(vcf = my_vcf)
vcf_filter_oneSNP(my_vcf)

}
\author{
Tomas Hrbek February 2022
}
