% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_filter_quality.R
\name{vcf_filter_quality}
\alias{vcf_filter_quality}
\title{vcf_filter_quality}
\usage{
vcf_filter_quality(vcf, qual = 20)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
subsetted vcfR object
}
\description{
remove loci below quality threshold from vcfR format data
}
\details{
This function removes loci below quality threshold from the vcfR object
}
\examples{
vcf_filter_quality(vcf = my_vcf, qual = 20)
vcf_filter_quality(my_vcf)

}
\author{
Tomas Hrbek February 2022
}
