% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_merge.R
\name{vcf_merge}
\alias{vcf_merge}
\title{vcf_merge}
\usage{
vcf_merge(vcf, vcf1)
}
\arguments{
\item{vcf}{-> vcfR object}

\item{vcf1}{-> vcfR object}
}
\value{
augmented vcfR object
}
\description{
merge two vcf; merge second vcf into first one
}
\details{
This function adds all individuals from a second vcfR object into a first vcfR object, returning new vcfR object
}
\examples{
vcf_merge(vcf = my_vcf, vcf1 = other_vcf)
vcf_merge(my_vcf, other_vcf)

}
\author{
Tomas Hrbek October 2023
}
