% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_stats.R
\name{vcf_stats}
\alias{vcf_stats}
\title{vcf_stats}
\usage{
vcf_stats(vcf, res_path, project)
}
\arguments{
\item{vcf}{-> vcfR object}
}
\value{
table of statistics
}
\description{
calculates basic stats of each samples from vcfR format data

high heterozygosity is indicative of potential contamination

average read depth per individual

missing data per individual
}
\details{
This function calculates average read depth, heterozygosity
number of heterozygotes, number of reference and alternative homozygotes,
missing data and total number SNPs of each sample in an vcfR object
}
\examples{
vcf_stats(vcf = my_vcf)
vcf_stats(my_vcf)

}
\author{
Tomas Hrbek August 2024
}
